Extract interface from BuildMutatorConfigurationSpec
