+++ src/main/java/org/gradle/profiler/mutations/BuildMutatorConfigurator.java
@@ -7,41 +7,15 @@ import org.gradle.profiler.InvocationSettings;
 public interface BuildMutatorConfigurator {
     BuildMutator configure(String key, BuildMutatorConfiguratorSpec spec);
 
-    class BuildMutatorConfiguratorSpec {
-        private final Config scenario;
-        private final String scenarioName;
-        private final InvocationSettings invocationSettings;
-        private final int warmupCount;
-        private final int buildCount;
+    interface BuildMutatorConfiguratorSpec {
+        Config getScenario();
 
-        public BuildMutatorConfiguratorSpec(Config scenario, String scenarioName, InvocationSettings invocationSettings, int warmupCount, int buildCount) {
-            this.scenario = scenario;
-            this.scenarioName = scenarioName;
-            this.invocationSettings = invocationSettings;
-            this.warmupCount = warmupCount;
-            this.buildCount = buildCount;
-        }
+        String getScenarioName();
 
-        public Config getScenario() {
-            return scenario;
-        }
+        int getWarmupCount();
 
-        public String getScenarioName() {
-            return scenarioName;
-        }
+        int getBuildCount();
 
-        public int getWarmupCount() {
-            return warmupCount;
-        }
-
-        public int getBuildCount() {
-            return buildCount;
-        }
-
-        public InvocationSettings getInvocationSettings() {
-            return invocationSettings;
-        }
-
+        InvocationSettings getInvocationSettings();
     }
-
 }
